Rename Edge classes to Edge, EdgeUndirected and EdgeDirected
