+++ src/Set/Edges.php
@@ -2,7 +2,7 @@
 
 namespace Graphp\Graph\Set;
 
-use Graphp\Graph\Edge\Base as Edge;
+use Graphp\Graph\Edge;
 use Graphp\Graph\Exception\InvalidArgumentException;
 use Graphp\Graph\Exception\OutOfBoundsException;
 use Graphp\Graph\Exception\UnderflowException;
