+++ src/Vertex.php
@@ -4,8 +4,6 @@ namespace Graphp\Graph;
 
 use Graphp\Graph\Attribute\AttributeAware;
 use Graphp\Graph\Attribute\AttributeBagReference;
-use Graphp\Graph\Edge\Base as Edge;
-use Graphp\Graph\Edge\Directed as EdgeDirected;
 use Graphp\Graph\Exception\BadMethodCallException;
 use Graphp\Graph\Exception\InvalidArgumentException;
 use Graphp\Graph\Set\Edges;
